Initial Android UI Commit
This is the initial commit for the Android UI code for Chassip. If you
want to be able to run it, you'll need to do a few things before that
happens. It took me forever to get it set up. I'll put together a
procedure so anyone can do it. In the meantime you can ask me.

Currently, the app has a left navigation drawer, a right one, and 5 tabs
that display a different color. The purpose of this is to demonstrate
that this type of interface makes sense for the larger scheme and is
easy for someone to use. Thus far, this has been great but obviously
there can always be improvements. I'll send out screenshots if you're
interested in seeing what it looks like right now (or I can send you the
apk).

My next steps are to lay the framework for user accounts, messages,
chats, and whispers/side-convos.

The chassip-parent is structured like this:
chassip-parent/
actionbarsherlock-parent/
actionbarsherlock/
pom.xml
slidingmenu-parent/
slidingmenu/
pom.xml
chassip/
pom.xml
pom.xml
myproject-api/
myproject-app/
pom.xml

-gpl